requirements.txt
anthropic
openai
numpy
pillow
python-dotenv